Split some files, localize language and region options
